fix: run protorev distro on epoch
