+++ x/protorev/keeper/hooks_test.go
@@ -2,10 +2,7 @@ package keeper_test
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
-
 	"github.com/osmosis-labs/osmosis/osmomath"
-	"github.com/osmosis-labs/osmosis/v24/app/apptesting"
 	"github.com/osmosis-labs/osmosis/v24/x/gamm/pool-models/balancer"
 	"github.com/osmosis-labs/osmosis/v24/x/gamm/pool-models/stableswap"
 	poolmanagertypes "github.com/osmosis-labs/osmosis/v24/x/poolmanager/types"
@@ -847,113 +844,6 @@ func (s *KeeperTestSuite) TestCompareAndStorePool() {
 				s.Require().NoError(err)
 				s.Require().Equal(expectedStoredPoolId, storedPoolId)
 			}
-		})
-	}
-}
-
-func (s *KeeperTestSuite) TestAfterEpochEnd() {
-	tests := []struct {
-		name       string
-		arbProfits sdk.Coins
-	}{
-		{
-			name:       "osmo denom only",
-			arbProfits: sdk.NewCoins(sdk.NewCoin("uosmo", osmomath.NewInt(100000000))),
-		},
-		{
-			name: "osmo denom and another base denom",
-			arbProfits: sdk.NewCoins(sdk.NewCoin("uosmo", osmomath.NewInt(100000000)),
-				sdk.NewCoin("juno", osmomath.NewInt(100000000))),
-		},
-		{
-			name: "osmo denom, another base denom, and a non base denom",
-			arbProfits: sdk.NewCoins(sdk.NewCoin("uosmo", osmomath.NewInt(100000000)),
-				sdk.NewCoin("juno", osmomath.NewInt(100000000)),
-				sdk.NewCoin("eth", osmomath.NewInt(100000000))),
-		},
-		{
-			name:       "no profits",
-			arbProfits: sdk.Coins{},
-		},
-	}
-
-	for _, tc := range tests {
-		s.Run(tc.name, func() {
-			s.SetupPoolsTest()
-
-			// Set base denoms
-			baseDenoms := []types.BaseDenom{
-				{
-					Denom:    types.OsmosisDenomination,
-					StepSize: osmomath.NewInt(1_000_000),
-				},
-				{
-					Denom:    "juno",
-					StepSize: osmomath.NewInt(1_000_000),
-				},
-			}
-			s.App.ProtoRevKeeper.SetBaseDenoms(s.Ctx, baseDenoms)
-
-			// Set protorev developer account
-			devAccount := apptesting.CreateRandomAccounts(1)[0]
-			s.App.ProtoRevKeeper.SetDeveloperAccount(s.Ctx, devAccount)
-
-			err := s.App.BankKeeper.MintCoins(s.Ctx, types.ModuleName, tc.arbProfits)
-			s.Require().NoError(err)
-
-			communityPoolBalancePre := s.App.BankKeeper.GetAllBalances(s.Ctx, s.App.AccountKeeper.GetModuleAddress(distrtypes.ModuleName))
-
-			// System under test
-			err = s.App.ProtoRevKeeper.AfterEpochEnd(s.Ctx, "day", 1)
-
-			expectedDevProfit := sdk.Coins{}
-			expectedOsmoBurn := sdk.Coins{}
-			arbProfitsBaseDenoms := sdk.Coins{}
-			arbProfitsNonBaseDenoms := sdk.Coins{}
-
-			// Split the profits into base and non base denoms
-			for _, coin := range tc.arbProfits {
-				isBaseDenom := false
-				for _, baseDenom := range baseDenoms {
-					if coin.Denom == baseDenom.Denom {
-						isBaseDenom = true
-						break
-					}
-				}
-				if isBaseDenom {
-					arbProfitsBaseDenoms = append(arbProfitsBaseDenoms, coin)
-				} else {
-					arbProfitsNonBaseDenoms = append(arbProfitsNonBaseDenoms, coin)
-				}
-			}
-			profitSplit := types.ProfitSplitPhase1
-			for _, arbProfit := range arbProfitsBaseDenoms {
-				devProfitAmount := arbProfit.Amount.MulRaw(profitSplit).QuoRaw(100)
-				expectedDevProfit = append(expectedDevProfit, sdk.NewCoin(arbProfit.Denom, devProfitAmount))
-			}
-
-			// Get the developer account balance
-			devAccountBalance := s.App.BankKeeper.GetAllBalances(s.Ctx, devAccount)
-			s.Require().Equal(expectedDevProfit, devAccountBalance)
-
-			// Get the burn address balance
-			burnAddressBalance := s.App.BankKeeper.GetAllBalances(s.Ctx, types.DefaultNullAddress)
-			if arbProfitsBaseDenoms.AmountOf(types.OsmosisDenomination).IsPositive() {
-				expectedOsmoBurn = sdk.NewCoins(sdk.NewCoin(types.OsmosisDenomination, arbProfitsBaseDenoms.AmountOf(types.OsmosisDenomination).Sub(expectedDevProfit.AmountOf(types.OsmosisDenomination))))
-				s.Require().Equal(expectedOsmoBurn, burnAddressBalance)
-			} else {
-				s.Require().Equal(sdk.Coins{}, burnAddressBalance)
-			}
-
-			// Get the community pool balance
-			communityPoolBalancePost := s.App.BankKeeper.GetAllBalances(s.Ctx, s.App.AccountKeeper.GetModuleAddress(distrtypes.ModuleName))
-			actualCommunityPool := communityPoolBalancePost.Sub(communityPoolBalancePre...)
-			expectedCommunityPool := arbProfitsBaseDenoms.Sub(expectedDevProfit...).Sub(expectedOsmoBurn...)
-			s.Require().Equal(expectedCommunityPool, actualCommunityPool)
-
-			// The protorev module account should only contain the non base denoms if there are any
-			protorevModuleAccount := s.App.BankKeeper.GetAllBalances(s.Ctx, s.App.AccountKeeper.GetModuleAddress(types.ModuleName))
-			s.Require().Equal(arbProfitsNonBaseDenoms, protorevModuleAccount)
 		})
 	}
 }
